refactor: merge deplicate logics
